Moved top-level classes into core.
