+++ include/nnlib/core/storage.h
