+++ include/nnlib/opt/optimizer.h
@@ -1,7 +1,7 @@
 #ifndef OPTIMIZER_H
 #define OPTIMIZER_H
 
-#include "../tensor.h"
+#include "../core/tensor.h"
 
 namespace nnlib
 {
