+++ include/nnlib/util/args.h
@@ -1,7 +1,7 @@
 #ifndef ARGS_H
 #define ARGS_H
 
-#include "../error.h"
+#include "../core/error.h"
 #include <unordered_map>
 #include <map>
 #include <vector>
